+++ ChangeLog.md
@@ -1,3 +1,7 @@
+# (unreleased)
+
+* Make the description more precise.
+
 # 0.0.3 (2014-03-26)
 
 * Add a more verbose description of the library.